+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Note: CHANGELOG was created in version 1.1.0, so change logs in versions that pr
 ### Added
 
 - Add Safe Mode(It is enabled by default and cannot be changed dynamically) in config.
+- Add "Math.mathfunc" to get a math function by name.
 
 ### Changed